added important swapfile commands

+++ 0-preinstall.sh
@@ -58,23 +58,21 @@ echo "keyserver hkp://keyserver.ubuntu.com" >> /mnt/etc/pacman.d/gnupg/gpg.conf
 echo "DISK=$DISK" >> "$SCRIPT_DIR/install.conf"
 cp -R "${SCRIPT_DIR}" /mnt/root
 cp /etc/pacman.d/mirrorlist /mnt/etc/pacman.d/mirrorlist
-echo -ne "
---------------------------------------------------------------------
-                 Checking for low memory systems <8G
---------------------------------------------------------------------
-"
 TOTALMEM=$(cat /proc/meminfo | grep -i 'memtotal' | grep -o '[[:digit:]]*')
-if [[ $TOTALMEM -lt 8000000 ]]; then
-    # Put swap into the actual system, not into RAM disk, otherwise there is no point in it, it'll cache RAM into RAM. So, /mnt/ everything.
+if [[ $TOTALMEM -lt 8000000 ]] && ! grep -q swapfile /mnt/etc/fstab; then
+    echo
+    echo "--------------------------------------------------------------------"
+    echo "                         Creating Swap File                         "
+    echo "--------------------------------------------------------------------"
     mkdir /mnt/opt/swap # make a dir that we can apply NOCOW to to make it btrfs-friendly.
+    truncate -s 0 /mnt/opt/swap
     chattr +C /mnt/opt/swap # apply NOCOW, btrfs needs that.
+    btrfs property set /mnt/opt/swap compression none
     dd if=/dev/zero of=/mnt/opt/swap/swapfile bs=1M count=2048 status=progress
-    chmod 600 /mnt/opt/swap/swapfile # set permissions.
-    chown root /mnt/opt/swap/swapfile
+    chmod 600 /mnt/opt/swap/swapfile
     mkswap /mnt/opt/swap/swapfile
     swapon /mnt/opt/swap/swapfile
-    # The line below is written to /mnt/ but doesn't contain /mnt/, since it's just / for the system itself.
-    echo "/opt/swap/swapfile	none	swap	sw	0	0" >> /mnt/etc/fstab # Add swap to fstab, so it KEEPS working after installation.
+    echo "/opt/swap/swapfile	none	swap	sw	0	0" >> /mnt/etc/fstab
 fi
 echo -ne "
 --------------------------------------------------------------------
